qkd-webapp-main/qkd_webapp/requirements.txt
fastapi
uvicorn
python-multipart
qiskit
qiskit-superstaq
numpy
pydantic
typing_extensions
matplotlib
python-socketio
python-dotenv
pylatexenc